+++ logi6816/src/name/bizna/emu65816/NativeModeInterrupts.java
@@ -2,19 +2,19 @@ package name.bizna.emu65816;
 
 public class NativeModeInterrupts
 {
-  public short coProcessorEnable;
-  public short brk;
-  public short abort;
-  public short nonMaskableInterrupt;
-  public short reset;
-  public short interruptRequest;
+  public int coProcessorEnable;
+  public int brk;
+  public int abort;
+  public int nonMaskableInterrupt;
+  public int reset;
+  public int interruptRequest;
 
-  public NativeModeInterrupts(short coProcessorEnable,
-                              short brk,
-                              short abort,
-                              short nonMaskableInterrupt,
-                              short reset,
-                              short interruptRequest)
+  public NativeModeInterrupts(int coProcessorEnable,
+                              int brk,
+                              int abort,
+                              int nonMaskableInterrupt,
+                              int reset,
+                              int interruptRequest)
   {
     this.coProcessorEnable = coProcessorEnable;
     this.brk = brk;
