+++ logi6816/src/name/bizna/emu65816/opcode/OpCode_PLX.java
@@ -18,7 +18,7 @@ public class OpCode_PLX
   {
     if (cpu.indexIs8BitWide())
     {
-      byte value = cpu.getStack().pull8Bit(cpu);
+      int value = cpu.getStack().pull8Bit(cpu);
       cpu.setX(Binary.setLower8BitsOf16BitsValue(cpu.getX(), value));
       cpu.getCpuStatus().updateSignAndZeroFlagFrom8BitValue(value);
       cpu.addToProgramAddressAndCycles(1, 4);
