+++ logi6816/src/name/bizna/emu65816/opcode/OpCode_TYA.java
@@ -17,19 +17,19 @@ public class OpCode_TYA
   {
     if (cpu.accumulatorIs8BitWide() && cpu.indexIs8BitWide())
     {
-      byte value = Binary.lower8BitsOf(cpu.getY());
+      int value = Binary.lower8BitsOf(cpu.getY());
       cpu.setA(Binary.setLower8BitsOf16BitsValue(cpu.getA(), value));
       cpu.getCpuStatus().updateSignAndZeroFlagFrom8BitValue(value);
     }
     else if (cpu.accumulatorIs8BitWide() && cpu.indexIs16BitWide())
     {
-      byte value = Binary.lower8BitsOf(cpu.getY());
+      int value = Binary.lower8BitsOf(cpu.getY());
       cpu.setA(Binary.setLower8BitsOf16BitsValue(cpu.getA(), value));
       cpu.getCpuStatus().updateSignAndZeroFlagFrom8BitValue(value);
     }
     else if (cpu.accumulatorIs16BitWide() && cpu.indexIs8BitWide())
     {
-      byte value = Binary.lower8BitsOf(cpu.getY());
+      int value = Binary.lower8BitsOf(cpu.getY());
       cpu.setA(value);
       cpu.getCpuStatus().updateSignAndZeroFlagFrom8BitValue(value);
     }
@@ -38,7 +38,6 @@ public class OpCode_TYA
       cpu.setA(cpu.getY());
       cpu.getCpuStatus().updateSignAndZeroFlagFrom16BitValue(cpu.getA());
     }
-
     cpu.addToProgramAddressAndCycles(1, 2);
   }
 }
